Structurize performances
